+++ packages/netlify-cms-backend-git-gateway/src/implementation.js
@@ -63,7 +63,7 @@ export default class GitGateway {
     const backendTypeMatches = this.gatewayUrl.match(backendTypeRegex);
     if (backendTypeMatches) {
       this.backendType = backendTypeMatches[1];
-      this.gatewayUrl = this.gatewayUrl.replace(backendTypeRegex, '/');
+      this.gatewayUrl = this.gatewayUrl.replace(backendTypeRegex, '');
     } else {
       this.backendType = null;
     }
@@ -131,7 +131,7 @@ export default class GitGateway {
       }
 
       const userData = {
-        name: user.user_metadata.name || user.email.split('@').shift(),
+        name: user.user_metadata.full_name || user.email.split('@').shift(),
         email: user.email,
         avatar_url: user.user_metadata.avatar_url,
         metadata: user.user_metadata,