+++ packages/netlify-cms-backend-gitlab/src/AuthenticationPage.js
@@ -17,7 +17,7 @@ export default class GitLabAuthenticationPage extends React.Component {
     siteId: PropTypes.string,
     authEndpoint: PropTypes.string,
     config: ImmutablePropTypes.map,
-    clearHash: PropTypes.function,
+    clearHash: PropTypes.func,
   };
 
   state = {};
@@ -63,12 +63,13 @@ export default class GitLabAuthenticationPage extends React.Component {
   };
 
   render() {
-    const { inProgress } = this.props;
+    const { inProgress, config } = this.props;
     return (
       <AuthenticationPage
         onLogin={this.handleLogin}
         loginDisabled={inProgress}
         loginErrorMessage={this.state.loginError}
+        logoUrl={config.get('logo_url')}
         renderButtonContent={() => (
           <React.Fragment>
             <LoginButtonIcon type="gitlab" /> {inProgress ? 'Logging in...' : 'Login with GitLab'}