+++ packages/netlify-cms-core/src/components/Collection/Sidebar.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import styled, { css } from 'react-emotion';
+import { translate } from 'react-polyglot';
 import { NavLink } from 'react-router-dom';
 import { Icon, components, colors, colorsRaw, lengths } from 'netlify-cms-ui-default';
 import { searchCollections } from 'Actions/collections';
@@ -34,7 +35,7 @@ const SidebarHeading = styled.h2`
 const SearchContainer = styled.div`
   display: flex;
   align-items: center;
-  margin: 0 8px;
+  margin: 0 12px;
   position: relative;
 
   ${Icon} {
@@ -64,6 +65,11 @@ const SearchInput = styled.input`
   }
 `;
 
+const SidebarNavList = styled.ul`
+  margin: 16px 0 0;
+  list-style: none;
+`;
+
 const SidebarNavLink = styled(NavLink)`
   display: flex;
   font-size: 14px;
@@ -72,25 +78,24 @@ const SidebarNavLink = styled(NavLink)`
   padding: 8px 12px;
   border-left: 2px solid #fff;
 
+  ${Icon} {
+    margin-right: 8px;
+  }
+
   ${props => css`
     &:hover,
     &:active,
     &.${props.activeClassName} {
       ${styles.sidebarNavLinkActive};
     }
-  `} &:first-of-type {
-    margin-top: 16px;
-  }
-
-  ${Icon} {
-    margin-right: 8px;
-  }
+  `};
 `;
 
-export default class Sidebar extends React.Component {
+class Sidebar extends React.Component {
   static propTypes = {
     collections: ImmutablePropTypes.orderedMap.isRequired,
     searchTerm: PropTypes.string,
+    t: PropTypes.func.isRequired,
   };
 
   static defaultProps = {
@@ -102,35 +107,35 @@ export default class Sidebar extends React.Component {
   renderLink = collection => {
     const collectionName = collection.get('name');
     return (
-      <SidebarNavLink
-        key={collectionName}
-        to={`/collections/${collectionName}`}
-        activeClassName="sidebar-active"
-      >
-        <Icon type="write" />
-        {collection.get('label')}
-      </SidebarNavLink>
+      <li key={collectionName}>
+        <SidebarNavLink to={`/collections/${collectionName}`} activeClassName="sidebar-active">
+          <Icon type="write" />
+          {collection.get('label')}
+        </SidebarNavLink>
+      </li>
     );
   };
 
   render() {
-    const { collections } = this.props;
+    const { collections, t } = this.props;
     const { query } = this.state;
 
     return (
       <SidebarContainer>
-        <SidebarHeading>Collections</SidebarHeading>
+        <SidebarHeading>{t('collection.sidebar.collections')}</SidebarHeading>
         <SearchContainer>
           <Icon type="search" size="small" />
           <SearchInput
             onChange={e => this.setState({ query: e.target.value })}
             onKeyDown={e => e.key === 'Enter' && searchCollections(query)}
-            placeholder="Search all"
+            placeholder={t('collection.sidebar.searchAll')}
             value={query}
           />
         </SearchContainer>
-        {collections.toList().map(this.renderLink)}
+        <SidebarNavList>{collections.toList().map(this.renderLink)}</SidebarNavList>
       </SidebarContainer>
     );
   }
 }
+
+export default translate()(Sidebar);