+++ packages/netlify-cms-widget-date/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/compare/netlify-cms-widget-date@2.0.6...netlify-cms-widget-date@2.1.0) (2018-11-29)
+
+
+### Bug Fixes
+
+* **a11y:** correct label "for" references to fields ([#1904](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/issues/1904)) ([955f94f](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/commit/955f94f))
+
+
+### Features
+
+* **widget-date:** add input display formatting ([#1739](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/issues/1739)) ([855efd8](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/commit/855efd8))
+
+
+
+
+
+## [2.0.6](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/compare/netlify-cms-widget-date@2.0.5...netlify-cms-widget-date@2.0.6) (2018-11-12)
+
+
+### Bug Fixes
+
+* **widget-date:** fix default value for nested date widget ([#1859](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/issues/1859)) ([d8f8887](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/commit/d8f8887))
+
+
+
+
+
 <a name="2.0.5"></a>
 ## [2.0.5](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-date/compare/netlify-cms-widget-date@2.0.4...netlify-cms-widget-date@2.0.5) (2018-08-24)